move ChangeUserRequest
