+++ graylog2-server/src/main/java/org/graylog2/users/UserManagementServiceImpl.java
@@ -23,13 +23,12 @@ import org.graylog2.Configuration;
 import org.graylog2.database.MongoConnection;
 import org.graylog2.plugin.database.ValidationException;
 import org.graylog2.plugin.database.users.User;
-import org.graylog2.rest.models.users.requests.ChangeUserRequest;
+import org.graylog2.shared.users.ChangeUserRequest;
 import org.graylog2.security.AccessTokenService;
 import org.graylog2.security.InMemoryRolePermissionResolver;
 import org.graylog2.shared.users.UserManagementService;
 
 import javax.inject.Inject;
-import javax.validation.Valid;
 
 public class UserManagementServiceImpl extends UserServiceImpl implements UserManagementService {
